requirements.txt
streamlit
boto3
pillow
streamlit_autorefresh
